Update for 6.8.8-3

+++ scripts/build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
-PVE_COMMIT=4cab886f26f9c8638593b8c553996c97ca9acc21
-PVE_VERSION=6.8.4-2
+PVE_COMMIT=69b55c504cab6549ee4ddcf8cef87c32699fbdb0
+PVE_VERSION=6.8.8-3
+NOTES="UNetLab Kernel for PVE 6.8.4-2\nOriginal commit:\n\n- Title: update ABI file for ${PVE_VERSION}-pve\n- Hash: \`${PVE_COMMIT}\`\n- Link: [git.proxmox.com](https://git.proxmox.com/?p=pve-kernel.git;a=commit;h=${PVE_COMMIT})"
 
 set -e
 
@@ -16,7 +17,10 @@ fi
 apt-get install -y git build-essential dh-make dh-python sphinx-common asciidoc-base bison dwarves flex libdw-dev libelf-dev libiberty-dev libnuma-dev libslang2-dev libssl-dev lintian lz4 python3-dev xmlto zlib1g-dev
 
 # Removing conflicting packages
-apt-get purge -y proxmox-headers*
+HEADERS=$(dpkg -l | egrep "proxmox-headers" | wc -l)
+if [ $HEADERS -ne 0 ]; then
+    apt-get purge -y proxmox-headers*
+fi
 
 # Cloning UNetLab patch
 git clone https://github.com/dainok/unetlab-kernel /usr/src/unetlab-kernel
@@ -30,8 +34,16 @@ make clean
 make build-dir-fresh
 
 # Patching the kernel
-sed -i /usr/src/pve-kernel/Makefile -e "s/EXTRAVERSION=.*/EXTRAVERSION=-\$\(KREL\)-pve-unl/"
+sed -i /usr/src/pve-kernel/Makefile -e "s/^EXTRAVERSION=.*/EXTRAVERSION=-\$\(KREL\)-pve-unl/"
 patch -p1 < /usr/src/unetlab-kernel/patches/transparent-bridge.patch
 
 # Compile the kernel
 make
+
+# Tag
+git tag ${PVE_VERSION} -a
+git push origin --tags
+
+# Create release
+echo -e $NOTES
+# gh release create ${PVE_VERSION} --latest --target=master --notes "${NOTES}" /usr/src/pve-kernel/proxmox-headers-*unl_*.deb /usr/src/pve-kernel/proxmox-kernel-*unl-signed*.deb /usr/src/pve-kernel/proxmox-kernel-*unl_*.deb
